[PL] TES: Explicitly narrow int.
Fixes processlib\tes\TESLocalAssemblerInner-impl.h(234): warning C4244: 'argument': conversion from 'const __int64' to 'const unsigned int', possible loss of data

+++ ProcessLib/TES/TESLocalAssemblerInner-impl.h
@@ -226,8 +226,10 @@ void TESLocalAssemblerInner<Traits>::assembleIntegrationPoint(
 {
     preEachAssembleIntegrationPoint(integration_point, localX, sm);
 
-    auto const N = sm.dNdx.cols();  // number of integration points
-    auto const D = sm.dNdx.rows();  // global dimension: 1, 2 or 3
+    auto const N =
+        static_cast<unsigned>(sm.dNdx.cols());  // number of integration points
+    auto const D =
+        static_cast<unsigned>(sm.dNdx.rows());  // global dimension: 1, 2 or 3
 
     assert(N * NODAL_DOF == local_M.cols());
